Add files via upload
It contains two recursive methods, one to sum the first n Fibonacci values and the other to determine the gcd of the two inputted sums. The main method contains the inputs, is called the recursive method, and outputs the results formally. The first recursive method for summing Fibonacci values uses memorization to increase efficiency.